requirements.txt
numpy==1.21.1
pandas==1.3.1
requests-oauthlib==1.3.0
Werkzeug==2.0.2
Flask==2.0.2
demoji==1.1.0
ja-sentence-segmenter==0.0.2
python-dotenv==0.15.0
ndjson==0.3.1
